+++ app/src/main/java/elder/ly/mobile/ui/screens/personalinfo/PersonalInfoScreen.kt
@@ -1,8 +1,6 @@
-package elder.ly.mobile.ui.screens.addressinfo
+package elder.ly.mobile.ui.screens.personalinfo
 
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
@@ -11,15 +9,18 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
 import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.SpanStyle
 import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.withStyle
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
@@ -28,42 +29,28 @@ import elder.ly.mobile.ui.components.BackIconButton
 import elder.ly.mobile.ui.components.DefaultDropdownMenu
 import elder.ly.mobile.ui.components.DefaultTextInput
 import elder.ly.mobile.ui.components.NextButton
-import elder.ly.mobile.utils.CustomMaskTranformation
 import elder.ly.mobile.ui.theme.MobileTheme
+import elder.ly.mobile.utils.CustomMaskTranformation
 
 @Composable
-fun InformacoesPessoaisEndereco() {
-    val brazilStates = listOf(
-        "AC", "AL", "AP", "AM", "BA", "CE", "DF", "ES", "GO", "MA",
-        "MT", "MS", "MG", "PA", "PB", "PR", "PE", "PI", "RJ", "RN",
-        "RS", "RO", "RR", "SC", "SP", "SE", "TO"
-    )
-
-    var cep by remember {
+fun PersonalInfoScreen() {
+    var fullName by remember {
         mutableStateOf("")
     }
 
-    var street by remember {
+    var email by remember {
         mutableStateOf("")
     }
 
-    var number by remember {
+    var document by remember {
         mutableStateOf("")
     }
 
-    var complement by remember {
+    var birthDate by remember {
         mutableStateOf("")
     }
 
-    var state by remember {
-        mutableStateOf("")
-    }
-
-    var city by remember {
-        mutableStateOf("")
-    }
-
-    var district by remember {
+    var gender by remember {
         mutableStateOf("")
     }
 
@@ -75,12 +62,13 @@ fun InformacoesPessoaisEndereco() {
     ) {
         Row(
             modifier = Modifier
-                .fillMaxWidth(),
+                .fillMaxWidth()
+                .padding(12.dp),
             horizontalArrangement = Arrangement.SpaceAround,
             verticalAlignment = Alignment.CenterVertically,
         ) {
             BackIconButton()
-            
+
             Text(
                 text = buildAnnotatedString {
                     withStyle(
@@ -88,94 +76,78 @@ fun InformacoesPessoaisEndereco() {
                             fontWeight = FontWeight.Bold
                         )
                     ) {
-                        append("Endereço")
+                        append("Informações\nAdicionais")
                     }
                 },
-                fontSize = 36.sp
+                fontSize = 36.sp,
+                textAlign = TextAlign.Center,
+                lineHeight = 40.sp
             )
 
             Spacer(modifier = Modifier.width(56.dp))
         }
 
         DefaultTextInput(
-            label = "CEP",
-            placeholder = "01234-567",
+            label = "Nome Completo",
+            value = fullName,
+            changeValue = { newFullName : String ->
+                fullName = newFullName
+            }
+        )
+
+        DefaultTextInput(
+            label = "Email",
+            placeholder = "seuemail@gmail.com",
+            keyboardType = KeyboardType.Email,
+            value = email,
+            changeValue = { newEmail : String ->
+                email = newEmail
+            }
+        )
+
+        DefaultTextInput(
+            label = "Documento (CPF/CNPJ)",
+            placeholder = "12345678910",
             keyboardType = KeyboardType.Number,
-            mask = CustomMaskTranformation(mask = "#####-###"),
+            maxChar = 11,
+            value = document,
+            changeValue = { newDocument : String ->
+                document = newDocument
+            }
+        )
+
+        DefaultTextInput(
+            label = "Data de Nascimento",
+            placeholder = "23/06/1991",
+            keyboardType = KeyboardType.Number,
+            mask = CustomMaskTranformation(mask = "##/##/####"),
             maxChar = 8,
-            value = cep,
-            changeValue = { newCep : String ->
-                cep = newCep
-            }
-        )
-
-        DefaultTextInput(
-            label = "Logradouro",
-            placeholder = "Rua Haddock Lobo",
-            value = street,
-            changeValue = { newStreet : String ->
-                street = newStreet
-            }
-        )
-
-        DefaultTextInput(
-            label = "Número",
-            placeholder = "12",
-            keyboardType = KeyboardType.Number,
-            maxChar = 6,
-            value = number,
-            changeValue = { newNumber : String ->
-                number = newNumber
-            }
-        )
-
-        DefaultTextInput(
-            label = "Complemento",
-            placeholder = "Bloco A",
-            value = complement,
-            changeValue = { newComplememt : String ->
-                complement = newComplememt
+            value = birthDate,
+            changeValue = { newBirthDate : String ->
+                birthDate = newBirthDate
             },
         )
 
         DefaultDropdownMenu(
-            label = "Estado",
-            placeholder = "Selecione um Estado",
-            options = brazilStates,
-            value = state,
-            changeValue = { newState : String ->
-                state = newState
+            label = "Gênero",
+            placeholder = "Selecione um Gênero",
+            options = listOf("Masculino", "Feminino", "Prefiro não Informar"),
+            value = gender,
+            changeValue = { newGender : String ->
+                gender = newGender
             }
-        )
-
-        DefaultTextInput(
-            label = "Cidade",
-            placeholder = "São Paulo",
-            value = city,
-            changeValue = { newCity : String ->
-                city = newCity
-            },
-        )
-
-        DefaultTextInput(
-            label = "Bairro",
-            placeholder = "Consolação",
-            value = district,
-            changeValue = { newDistrict : String ->
-                district = newDistrict
-            },
         )
 
         Spacer(modifier = Modifier.weight(1f))
 
-        NextButton(label = "Salvar")
+        NextButton(label = "Avançar")
     }
 }
 
 @Preview(showBackground = true)
 @Composable
-fun InformacoesPessoaisEnderecoPreview() {
+private fun PersonalInfoScreenPreview() {
     MobileTheme {
-        InformacoesPessoaisEndereco()
+        PersonalInfoScreen()
     }
 }
